+++ stanza/error.go
@@ -261,6 +261,29 @@ type Error struct {
 	Text      map[string]string
 }
 
+// Is will be used by errors.Is when comparing errors.
+// It compares the condition and type fields.
+// If either is empty it is treated as a wildcard.
+// If both are empty the comparison is true if err is of type Error.
+//
+// For more information see the errors package.
+func (se Error) Is(target error) bool {
+	err, ok := target.(Error)
+	if !ok {
+		return false
+	}
+
+	switch {
+	case err.Type == "" && err.Condition == "":
+		return true
+	case err.Type == "":
+		return err.Condition == se.Condition
+	case err.Condition == "":
+		return err.Type == se.Type
+	}
+	return err.Condition == se.Condition && err.Type == se.Type
+}
+
 // Error satisfies the error interface by returning the condition.
 func (se Error) Error() string {
 	return string(se.Condition)